Fixes for installing on 2017-01-11-raspbian-jessie-lite.img on Raspberrry Pi 3B
Two fixes I needed to do to get it all to work.
First fix for bug https://bugs.launchpad.net/ubuntu/+source/cryptsetup/+bug/595648 as described in the fifth comment. The actual file was slightly different compared to what was described in the link, but exactly as described there I needed to comment out the if branch of the if else (make sure to remember to also comment out the fi):
```
#if [ -x /bin/plymouth ] && plymouth --ping; then
                        #       cryptkeyscript="plymouth ask-for-password --prompt"
                        #       # Plymouth will add a : if it is a non-graphical prompt
                        #       cryptkey="Please unlock disk $diskname"
                        #else
                                cryptkeyscript="/lib/cryptsetup/askpass"
                                cryptkey="Please unlock disk $diskname: "
                        #fi
```.

Second: My Pi would successfully run the initramfs, but not connect to the network. I needed to add the `ip=:::::eth0:dhcp` kernel option to `/boot/cmdline.txt` for it to work.